// lib/env.server.d.ts
export declare const envServer: {
    NEXT_PUBLIC_SUPABASE_URL: string;
    NEXT_PUBLIC_SUPABASE_ANON_KEY: string;
    SUPABASE_SERVICE_ROLE_KEY: string;
    OPENAI_API_KEY: string;
    MISTRAL_API_KEY: string;
    VOICE_DEBUG: boolean;
    STRIPE_SECRET_KEY: string;
    NODE_ENV: "development" | "test" | "production";
    PORT: number;
    CI: boolean;
    SUPABASE_URL?: string | undefined;
    SUPABASE_ANON_KEY?: string | undefined;
    OPENAI_BASE_URL?: string | undefined;
    ELEVENLABS_API_KEY?: string | undefined;
    ELEVENLABS_VOICE_ID?: string | undefined;
    STRIPE_WEBHOOK_SECRET?: string | undefined;
    STRIPE_PRICE_EARLY?: string | undefined;
    STRIPE_PRO_PRICE_ID?: string | undefined;
    STRIPE_EXPERT_PRICE_ID?: string | undefined;
    UPSTASH_REDIS_REST_URL?: string | undefined;
    UPSTASH_REDIS_REST_TOKEN?: string | undefined;
    REDIS_URL?: string | undefined;
    RESEND_API_KEY?: string | undefined;
    SLACK_WEBHOOK_URL?: string | undefined;
    IPQS_KEY?: string | undefined;
    ABSTRACT_KEY?: string | undefined;
    TURN_URL?: string | undefined;
    DATABASE_URL?: string | undefined;
    CRON_SECRET?: string | undefined;
    POSTHOG_API_KEY?: string | undefined;
    DEEPGRAM_API_KEY?: string | undefined;
    STORE?: string | undefined;
};